perf/monitor_dashboard/backend/requirements.txt
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
sqlalchemy>=2.0.0
sqlite3
requests>=2.31.0
python-dateutil>=2.8.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.15.0
schedule>=1.2.0
apscheduler>=3.10.0
tqdm>=4.65.0
colorama>=0.4.6
watchdog>=3.0.0
pyyaml>=6.0
jinja2>=3.1.0